# Conflicts:
#	src/main/java/com/example/controllers/CuentaBancariaController.java
#	src/main/java/com/example/models/CuentaBancaria.java

+++ src/main/java/com/example/repository/cuentaBancariaRepositorio.java
@@ -3,6 +3,6 @@ package com.example.repository;
 import com.example.models.CuentaBancaria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface cuentaBancariaRepositorio extends JpaRepository<CuentaBancaria, Integer> {
+public interface cuentaBancariaRepositorio extends JpaRepository<CuentaBancaria, Long> {
 }
